Breaking Change: Update to leverage build process using keys.json (#19)
* Updating the way app key is leveraged

This change is reflected in the build process and is meant to allow for multiple targets to be defined at the project level.

* Updating readme for more clarity and information

Primarily to include the new key structure but also create more overall clarity on use.

* Updating externally connectable

Error is present without the empty keys.

* Updating default build directory to _dev

* Adding Initial index file for local development

This is really just boilerplate (with some functionality).  It would be helpful to flesh this out a bit and add more utility to the locally served index.  Some ideas include:

1. Storybooks would be served here.
2. Style guides
3. Setup information
4. Server info
5. Devtools info

* Updating readme to indicate file location for keys