Fix/integrity fixes for production launch (#24)
* Adding saga race to recommendation request

* Updating dev tools to use standard remote-dev-tool

* Adding check to ensure readiness

- Checking Frame is loaded
- Checking script is injected

* Updating save actions to `takeEvery`

So unique saves do not cancel each other out

* Updating remote-dev-tools dependencies

Also adding removed-server to avoid needing an external connection

* Updating raised state to include focus and panel

Knowing when to hide the panel was a challenge with locally managed input focus.  This moves the action of closing with a timeout into sagas and adds input focus to the top of the locally managed state.

* Adding frame click signals

Added a signal on the frame as well as the page so it is clear when the frame loses focus.  This is to address the possibility of the input being focused and the user clicking outside the frame, which would result in no closePanel action to be fired.

* Updating composeEnhancer to commit realtime

This is a setting to force dev-tools in production.  It will throw an error if it is left in and there is not an open socket for it to connect to.

* Fixing tag error with stored tags

Added a basic check to return false for tags if no tags were available