Removing js injection on all sites
The way we traditionally injected content scripts into all sites caused some issue that were easily resolved my moving to an on-demand model.  Since we already had on-demand functionality in place to handle pages that existed prior to extension install, simply removing that script from the content_scripts in the manifest moves us to fully on-demand for all sites.