Update/Eliminate pre-emptive tab storage (#30)
Tabs would be marked as `ready` after being loaded.  This had the potential to cause high memory consumption on a user with a great many tabs open.  This is being switched to a less aggressive version where we only text readiness on a save request.  We drop tabs when they are closed.