Update/caching cpu usage (#37)
* Add check for system page as well as newtab

We don’t want to attempt to save a system page (chrome://*) at the moment.  This will throw an error so better to control the behavior on our end and send them to getpocket.com

* Add Swap tabs functionality for _tabs state

If a tab get’s replaced on chromes side we need to update state on our side to match.  This just adds the new tab and deletes the old key.

* Cleaning up key_shortcuts as they are managed now

We use native management for keyboard shortcuts to reduce overhead and improve security

* Adjusting how tabs status is set

The old method rewrote the entire state of tabs each tab shift.  This will only operate on two tabs per shift.

* Removing save hashes

We will base our saves of tabs to avoid unwarranted save collections. For the life of the tab we will have a single save available.  This eliminates the need for the save hash and avoid the problem of exponential growth of the save objects.

* Updating panel visual interactions

This lets the panel slide in to view as opposed to just appearing.  This takes slightly longer but is perceived much faster.

The reason this is included in this PR is because it make testing a bit clearer for speed in high CPU loads

* Adding Extension pages to isSystemPage check

* Updating props to the toolbar

Passing through the current tab as opposed to current tab values.

* Fixing spelling error

* Update Tab state to reset when closed

* Adding check for used tags existing

* Fixing hover events for animated panel

* Adjusting visuals on animation

- Increasing force of spring
- Adding persevere-3d for text clarity

* Updating TabId in favor of SaveHash

* Updating panels to close and unload on switching

* Adding Feed  check on recommendation header

* Updating Recommendation animation settings

Slowed it down a bit so it wasn’t so jumpy with the slide in animation on save

* Fixed passing wrong prop for typeahead

* Cleaning up stray comments