Adding TabActive state management
This cleans up the active tab management. At this time we simply unload the frame when we switch away from the tab or window.  There is room here to make this a bit smarter and elegant.  This is the bare functionality to get things solid.

This also moves the MESSAGE section a bit lower for clarity