Moving Tagging and Suggestions
Suggestions are now a part of the tagging module. They can be omitted safely and it will reduce state intersections.

Toolbar now takes a child element (Tagging in this case, but could be anything).  This lets us more easily set up and maintian a seperate tagging container and keep the tagging state seperated.