chore(cleanup): adjusting code format
Ideally we will be able to automate this in the future since it can be highly contentious.  The goal here is to make sure code is readable at the same widths as the rest of the code.  In the JS we auto-wrap at a specific width, which has the added benefit of encouraging concise and discrete code.